+++ client/src/components/ui/List.jsx
@@ -1,0 +1,50 @@
+import React from "react";
+import { useSelector } from "react-redux";
+import CardItem from './CardItem';
+
+const List = ({ id }) => {
+  const list = useSelector(state => state.lists.find(list => list._id === id));
+  const cards = useSelector(state => state.cards.filter(card => card.listId === id));
+  console.log(cards);
+
+  return (
+    <div className="list-wrapper">
+      <div className="list-background">
+        <div className="list">
+          <a className="more-icon sm-icon" href=""></a>
+          <div>
+            <p className="list-title">{list.title}</p>
+          </div>
+          <div className="add-dropdown add-top">
+            <div className="card"></div>
+            <a className="button">Add</a>
+            <i className="x-icon icon"></i>
+            <div className="add-options">
+              <span>...</span>
+            </div>
+          </div>
+          <div id="cards-container" data-id={`list-${id}-cards`}>
+            {cards.map(card => <CardItem key={card._id} id={card._id} />)}
+          </div>
+          <div className="add-dropdown add-bottom">
+            <div className="card">
+              <div className="card-info"></div>
+              <textarea name="add-card"></textarea>
+              <div className="members"></div>
+            </div>
+            <a className="button">Add</a>
+            <i className="x-icon icon"></i>
+            <div className="add-options">
+              <span>...</span>
+            </div>
+          </div>
+          <div className="add-card-toggle" data-position="bottom">
+            Add a card...
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default List;
