+++ client/src/constants/ActionTypes.js
@@ -2,3 +2,4 @@ export const FETCH_BOARDS_REQUEST = "FETCH_BOARDS_REQUEST";
 export const FETCH_BOARDS_SUCCESS = "FETCH_BOARDS_SUCCESS";
 export const CREATE_BOARD_REQUEST = "CREATE_BOARD_REQUEST";
 export const CREATE_BOARD_SUCCESS = "CREATE_BOARD_SUCCESS";
+export const BOARD_FETCHED = "BOARD_FETCHED";
