+++ client/src/lib/ApiClient.js
@@ -32,6 +32,13 @@ const apiClient = {
       .then(unwrapData)
       .then(callback)
       .catch(logError);
+  },
+  getBoard: function(id, callback) {
+    return axios
+      .get(`${routes.BOARDS_INDEX_URL}/${id}`)
+      .then(unwrapData)
+      .then(callback)
+      .catch(logError);
   }
 };
