+++ client/src/reducers/rootReducer.js
@@ -1,6 +1,8 @@
 import { combineReducers } from "redux";
 import boards from "./boards";
+import lists from './lists';
+import cards from './cards';
 
-const rootReducer = combineReducers({ boards });
+const rootReducer = combineReducers({ boards, lists, cards });
 
 export default rootReducer;
