+++ server/routes/api.js
@@ -1,6 +1,8 @@
 const express = require("express");
 const router = express.Router();
 const boardsController = require("../controllers/boardsController");
+const listsController = require("../controllers/listsController");
+const cardsController = require("../controllers/cardsController");
 const {
   validateBoard,
   validateCard,
@@ -13,8 +15,8 @@ router.get("/boards", boardsController.getBoards);
 
 router.post("/boards", validateBoard, boardsController.createBoard);
 
-router.post("/lists", validateList, listsController.createList);
+// router.post("/lists", validateList, listsController.createList);
 
-router.post("/cards", validateCard, cardsController.createCard);
+// router.post("/cards", validateCard, cardsController.createCard);
 
 module.exports = router;
